Move QA tasks to DPG
