+++ Common/TableProducer/PID/pidTPC.cxx
@@ -18,22 +18,20 @@
 ///         Only the tables for the mass hypotheses requested are filled, the others are sent empty.
 ///
 
+// ROOT includes
 #include "TFile.h"
 
 // O2 includes
 #include "Framework/AnalysisTask.h"
-#include "Framework/HistogramRegistry.h"
 #include "ReconstructionDataFormats/Track.h"
 #include <CCDB/BasicCCDBManager.h>
 #include "Common/Core/PID/PIDResponse.h"
 #include "Common/Core/PID/TPCPIDResponse.h"
-#include "Common/DataModel/TrackSelectionTables.h"
 #include "Framework/AnalysisDataModel.h"
 #include "Common/DataModel/Multiplicity.h"
 #include "TableHelper.h"
-#include "Common/DataModel/EventSelection.h"
-#include "Framework/StaticFor.h"
 #include "Common/TableProducer/PID/pidTPCML.h"
+#include "DPG/Tasks/qaPIDTPC.h"
 
 using namespace o2;
 using namespace o2::framework;
@@ -248,176 +246,6 @@ struct tpcPid {
       makeTable(pidAl, tablePIDAl, o2::track::PID::Alpha);
 
       count_tracks++;
-    }
-  }
-};
-
-/// Task to produce the TPC QA plots
-struct tpcPidQa {
-  static constexpr int Np = 9;
-  static constexpr const char* pT[Np] = {"e", "#mu", "#pi", "K", "p", "d", "t", "^{3}He", "#alpha"};
-  static constexpr std::string_view hnsigma[Np] = {"nsigma/El", "nsigma/Mu", "nsigma/Pi",
-                                                   "nsigma/Ka", "nsigma/Pr", "nsigma/De",
-                                                   "nsigma/Tr", "nsigma/He", "nsigma/Al"};
-  static constexpr std::string_view hnsigmapt[Np] = {"nsigmapt/El", "nsigmapt/Mu", "nsigmapt/Pi",
-                                                     "nsigmapt/Ka", "nsigmapt/Pr", "nsigmapt/De",
-                                                     "nsigmapt/Tr", "nsigmapt/He", "nsigmapt/Al"};
-  static constexpr std::string_view hnsigmapospt[Np] = {"nsigmapospt/El", "nsigmapospt/Mu", "nsigmapospt/Pi",
-                                                        "nsigmapospt/Ka", "nsigmapospt/Pr", "nsigmapospt/De",
-                                                        "nsigmapospt/Tr", "nsigmapospt/He", "nsigmapospt/Al"};
-  static constexpr std::string_view hnsigmanegpt[Np] = {"nsigmanegpt/El", "nsigmanegpt/Mu", "nsigmanegpt/Pi",
-                                                        "nsigmanegpt/Ka", "nsigmanegpt/Pr", "nsigmanegpt/De",
-                                                        "nsigmanegpt/Tr", "nsigmanegpt/He", "nsigmanegpt/Al"};
-
-  HistogramRegistry histos{"Histos", {}, OutputObjHandlingPolicy::AnalysisObject};
-
-  Configurable<int> logAxis{"logAxis", 1, "Flag to use a log momentum axis"};
-  Configurable<int> nBinsP{"nBinsP", 3000, "Number of bins for the momentum"};
-  Configurable<float> minP{"minP", 0.01, "Minimum momentum in range"};
-  Configurable<float> maxP{"maxP", 20, "Maximum momentum in range"};
-  Configurable<int> nBinsNSigma{"nBinsNSigma", 200, "Number of bins for the NSigma"};
-  Configurable<float> minNSigma{"minNSigma", -10.f, "Minimum NSigma in range"};
-  Configurable<float> maxNSigma{"maxNSigma", 10.f, "Maximum NSigma in range"};
-  Configurable<int> applyEvSel{"applyEvSel", 2, "Flag to apply rapidity cut: 0 -> no event selection, 1 -> Run 2 event selection, 2 -> Run 3 event selection"};
-  Configurable<bool> applyTrackCut{"applyTrackCut", false, "Flag to apply standard track cuts"};
-  Configurable<bool> applyRapidityCut{"applyRapidityCut", false, "Flag to apply rapidity cut"};
-
-  template <uint8_t i>
-  void addParticleHistos(const AxisSpec& pAxis, const AxisSpec& ptAxis)
-  {
-    // NSigma
-    const char* axisTitle = Form("N_{#sigma}^{TPC}(%s)", pT[i]);
-    const AxisSpec nSigmaAxis{nBinsNSigma, minNSigma, maxNSigma, axisTitle};
-    histos.add(hnsigma[i].data(), axisTitle, kTH2F, {pAxis, nSigmaAxis});
-    histos.add(hnsigmapt[i].data(), axisTitle, kTH2F, {ptAxis, nSigmaAxis});
-    histos.add(hnsigmapospt[i].data(), axisTitle, kTH2F, {ptAxis, nSigmaAxis});
-    histos.add(hnsigmanegpt[i].data(), axisTitle, kTH2F, {ptAxis, nSigmaAxis});
-  }
-
-  void init(o2::framework::InitContext&)
-  {
-
-    const AxisSpec multAxis{1000, 0.f, 1000.f, "Track multiplicity"};
-    const AxisSpec vtxZAxis{100, -20, 20, "Vtx_{z} (cm)"};
-    const AxisSpec etaAxis{100, -2, 2, "#it{#eta}"};
-    const AxisSpec lAxis{100, 0, 500, "Track length (cm)"};
-    const AxisSpec pAxisPosNeg{nBinsP, -maxP, maxP, "Signed #it{p} (GeV/#it{c})"};
-    AxisSpec ptAxis{nBinsP, minP, maxP, "#it{p}_{T} (GeV/#it{c})"};
-    AxisSpec pAxis{nBinsP, minP, maxP, "#it{p} (GeV/#it{c})"};
-    if (logAxis) {
-      ptAxis.makeLogaritmic();
-      pAxis.makeLogaritmic();
-    }
-    const AxisSpec dedxAxis{1000, 0, 1000, "d#it{E}/d#it{x} A.U."};
-
-    // Event properties
-    auto h = histos.add<TH1>("event/evsel", "", kTH1F, {{10, 0.5, 10.5, "Ev. Sel."}});
-    h->GetXaxis()->SetBinLabel(1, "Events read");
-    h->GetXaxis()->SetBinLabel(2, "Passed ev. sel.");
-    h->GetXaxis()->SetBinLabel(3, "Passed mult.");
-    h->GetXaxis()->SetBinLabel(4, "Passed vtx Z");
-
-    histos.add("event/vertexz", "", kTH1F, {vtxZAxis});
-    h = histos.add<TH1>("event/particlehypo", "", kTH1F, {{10, 0, 10, "PID in tracking"}});
-    for (int i = 0; i < 9; i++) {
-      h->GetXaxis()->SetBinLabel(i + 1, PID::getName(i));
-    }
-    histos.add("event/trackmultiplicity", "", kTH1F, {multAxis});
-    histos.add("event/tpcsignal", "", kTH2F, {pAxis, dedxAxis});
-    histos.add("event/signedtpcsignal", "", kTH2F, {pAxisPosNeg, dedxAxis});
-    histos.add("event/eta", "", kTH1F, {etaAxis});
-    histos.add("event/length", "", kTH1F, {lAxis});
-    histos.add("event/pt", "", kTH1F, {ptAxis});
-    histos.add("event/p", "", kTH1F, {pAxis});
-
-    static_for<0, 8>([&](auto i) {
-      addParticleHistos<i>(pAxis, ptAxis);
-    });
-  }
-
-  template <o2::track::PID::ID id, typename T>
-  void fillParticleHistos(const T& t)
-  {
-    if (applyRapidityCut) {
-      const float y = TMath::ASinH(t.pt() / TMath::Sqrt(PID::getMass2(id) + t.pt() * t.pt()) * TMath::SinH(t.eta()));
-      if (abs(y) > 0.5) {
-        return;
-      }
-    }
-
-    // Fill histograms
-    const auto& nsigma = o2::aod::pidutils::tpcNSigma<id>(t);
-    histos.fill(HIST(hnsigma[id]), t.p(), nsigma);
-    histos.fill(HIST(hnsigmapt[id]), t.pt(), nsigma);
-    if (t.sign() > 0) {
-      histos.fill(HIST(hnsigmapospt[id]), t.pt(), nsigma);
-    } else {
-      histos.fill(HIST(hnsigmanegpt[id]), t.pt(), nsigma);
-    }
-  }
-
-  using Trks = soa::Join<aod::Tracks, aod::TracksExtra,
-                         aod::pidTPCEl, aod::pidTPCMu, aod::pidTPCPi,
-                         aod::pidTPCKa, aod::pidTPCPr, aod::pidTPCDe,
-                         aod::pidTPCTr, aod::pidTPCHe, aod::pidTPCAl,
-                         aod::TrackSelection>;
-  void process(soa::Join<aod::Collisions, aod::EvSels>::iterator const& collision,
-               Trks const& tracks)
-  {
-
-    histos.fill(HIST("event/evsel"), 1);
-    if (applyEvSel == 1) {
-      if (!collision.sel7()) {
-        return;
-      }
-    } else if (applyEvSel == 2) {
-      if (!collision.sel8()) {
-        return;
-      }
-    }
-
-    histos.fill(HIST("event/evsel"), 2);
-
-    // Computing Multiplicity first
-    float ntracks = 0;
-    for (auto t : tracks) {
-      if (applyTrackCut && !t.isGlobalTrack()) {
-        continue;
-      }
-      ntracks += 1;
-    }
-    // if (0 && ntracks < 1) {
-    //   return;
-    // }
-    histos.fill(HIST("event/evsel"), 3);
-    if (abs(collision.posZ()) > 10.f) {
-      return;
-    }
-    histos.fill(HIST("event/evsel"), 4);
-    histos.fill(HIST("event/vertexz"), collision.posZ());
-    histos.fill(HIST("event/trackmultiplicity"), ntracks);
-
-    for (auto t : tracks) {
-      if (applyTrackCut && !t.isGlobalTrack()) {
-        continue;
-      }
-      histos.fill(HIST("event/particlehypo"), t.pidForTracking());
-      histos.fill(HIST("event/tpcsignal"), t.tpcInnerParam(), t.tpcSignal());
-      histos.fill(HIST("event/signedtpcsignal"), t.tpcInnerParam() * t.sign(), t.tpcSignal());
-      histos.fill(HIST("event/eta"), t.eta());
-      histos.fill(HIST("event/length"), t.length());
-      histos.fill(HIST("event/pt"), t.pt());
-      histos.fill(HIST("event/p"), t.p());
-      //
-      fillParticleHistos<PID::Electron>(t);
-      fillParticleHistos<PID::Muon>(t);
-      fillParticleHistos<PID::Pion>(t);
-      fillParticleHistos<PID::Kaon>(t);
-      fillParticleHistos<PID::Proton>(t);
-      fillParticleHistos<PID::Deuteron>(t);
-      fillParticleHistos<PID::Triton>(t);
-      fillParticleHistos<PID::Helium3>(t);
-      fillParticleHistos<PID::Alpha>(t);
     }
   }
 };
